Fix issues from simplify review
- Add LatencyTracker.Max() for O(N) linear scan without allocation;
  use it in recordLatency hot path instead of Stats() which sorts
- Use ToInt64/ToInt16 at call sites instead of int64(ToFloat64(...))
  — these exported helpers were unused dead code before
- Hoist trackedThreadPools to package-level var to avoid per-poll
  map allocation in computeRejectionDeltas
- Clean up nodeHistories, prevIOSample, prevRejected when a node
  is pruned from knownNodes — prevents unbounded growth
- Remove redundant WHAT-not-WHY doc comments from convert.go

+++ internal/collector/health.go
@@ -51,8 +51,8 @@ func (c *HealthCollector) Collect(ctx context.Context, reg *cratedb.Registry, st
 			TableSchema:     cratedb.ToString(row[0]),
 			TableName:       cratedb.ToString(row[1]),
 			Health:          cratedb.ToString(row[2]),
-			MissingShards:   int64(cratedb.ToFloat64(row[3])),
-			UnderReplicated: int64(cratedb.ToFloat64(row[4])),
+			MissingShards:   cratedb.ToInt64(row[3]),
+			UnderReplicated: cratedb.ToInt64(row[4]),
 			Partition:       cratedb.ToString(row[5]),
 		}
 		health = append(health, h)
